useFeatureFlag for enable rewards flag
