+++ app/components/hooks/FeatureFlags/useFeatureFlag.ts
@@ -1,0 +1,12 @@
+import { useFeatureFlagOverride } from '../../../contexts/FeatureFlagOverrideContext';
+
+export enum FeatureFlagNames {
+  rewardsEnabled = 'rewardsEnabled',
+  rewardsEnableCardSpend = 'rewardsEnableCardSpend',
+  rewardsAnnouncementModalEnabled = 'rewardsAnnouncementModalEnabled',
+}
+
+export const useFeatureFlag = (key: FeatureFlagNames) => {
+  const { getFeatureFlag } = useFeatureFlagOverride();
+  return getFeatureFlag(key);
+};
